refactor: inline page composition in routes
