+++ src/routes/_protected/bookmarks/$id/edit.tsx
@@ -66,7 +66,7 @@ function isBookmarkNotFound(error: unknown): boolean {
 }
 
 /**
- * RouteComponent は編集画面の Screen 境界であり、Storybook の Route Story 起点でもある。
+ * RouteComponent は編集画面のページ境界であり、Storybook の Route Story 起点でもある。
  * params / search / loader / not-found / 画面固有リンク / navigation をここで閉じ、
  * Domain・DB・oRPC 実装詳細は注入された port の向こう側に置く。
  */
