+++ src/routes/_protected/tags/new.test.ts
@@ -6,9 +6,9 @@ import { describe, expect, test } from 'vitest'
 
 const dir = dirname(fileURLToPath(import.meta.url))
 
-describe('NewTagScreen', () => {
+describe('new tag route', () => {
   test('Error の class 名に依存しない', () => {
-    const source = readFileSync(join(dir, 'new-tag-screen.tsx'), 'utf8')
+    const source = readFileSync(join(dir, 'new.tsx'), 'utf8')
     expect(source).not.toContain('TagNameAlreadyExistsError')
     expect(source).not.toContain('error.name')
     expect(source).toContain('getCreateTagErrorMessage')
